requirements.txt
pika==0.9.8
requests
tornado
pyyaml
filemagic
Celery